rename layout files
