--- app/src/main/java/com/map08/shiftapp/pages/EmployeeHomeScreen.kt
@@ -1,30 +1,0 @@
-package com.map08.shiftapp.pages
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun EmployeeHomeScreen(){
-    Column(
-        modifier = Modifier.fillMaxSize().background(Color(0XFF1976D2)),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = "Home Page",
-            fontSize = 40.sp,
-            fontWeight = FontWeight.SemiBold,
-            color = Color.White
-        )
-
-    }
-}
